+++ src/report-renderer.js
@@ -4,6 +4,7 @@ import { renderTextReport } from './report-text.js';
 
 /**
  * @typedef {object} RenderOptions
+ * @property {string} datetimeFormat HTML report datetime display format.
  * @property {boolean} forceRefresh Whether HTML output should include the calculated refresh timer.
  * @property {string} format Output format.
  * @property {number | undefined} interval Optional regeneration interval in seconds.
@@ -23,6 +24,7 @@ export function renderReport(report, options) {
     }
     if (options.format === 'html') {
         return renderHtmlReport(report, {
+            datetimeFormat: options.datetimeFormat,
             refreshSeconds: options.forceRefresh
                 ? (options.interval ?? 0) - 2
                 : undefined,